Move code from `Misc.swift` to other files
Nobody loves files whose best description is “miscellaneous”.
