+++ CodeGeneration/Sources/generate-swiftsyntax/GenerateSwiftSyntax.swift
@@ -99,7 +99,6 @@ struct GenerateSwiftSyntax: ParsableCommand {
 
         // SwiftSyntax
         GeneratedFileSpec(swiftSyntaxGeneratedDir + ["Keyword.swift"], keywordFile),
-        GeneratedFileSpec(swiftSyntaxGeneratedDir + ["Misc.swift"], miscFile),
         GeneratedFileSpec(swiftSyntaxGeneratedDir + ["raw", "RawSyntaxNodes.swift"], rawSyntaxNodesFile),
         GeneratedFileSpec(swiftSyntaxGeneratedDir + ["raw", "RawSyntaxValidation.swift"], rawSyntaxValidationFile),
         GeneratedFileSpec(swiftSyntaxGeneratedDir + ["SyntaxAnyVisitor.swift"], syntaxAnyVisitorFile),
